requirements.txt
Django==3.2.6
Pillow==8.3.1
django-admin-sortable2==1.0
django-tinymce==3.3.0

